requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
SQLAlchemy==2.0.21
#psycopg2
python-dotenv==0.21.0
pytest
pytest-cov
